optimise imports, reformat code, in tests

+++ mes/src/test/java/org/grp2/OEETest.java
@@ -9,7 +9,6 @@ import java.util.Collections;
 import java.util.List;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 
 public class OEETest {
 
@@ -25,7 +24,7 @@ public class OEETest {
             Double stopTime = stopTimes.get(i);
             OEE oee = new OEE(batch, stopTime);
 
-            assertEquals((double)expectedOEE.get(i), (double)oee.getOEE(), 0);
+            assertEquals(expectedOEE.get(i), oee.getOEE(), 0);
         }
     }
